Locomotives can now be retrieved via CAN
Added zlib inflator and message parsing. Some refactoring
